% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_web_functions.R
\name{capture_yml}
\alias{capture_yml}
\title{Capture yaml lines to write}
\usage{
capture_yml(yml)
}
\arguments{
\item{yml}{yml object (\code{\link[ymlthis]{as_yml}})}
}
\value{
captured output (lines) without color attributes of the provided ylm
}
\description{
Capture yaml lines to write
}
