% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{collect_metadata}
\alias{collect_metadata}
\title{Collect metadata from yml files}
\usage{
collect_metadata(con, .dry = TRUE)
}
\arguments{
\item{con}{Connection to the metadata database (DBI or pool object)}

\item{.dry}{Logical indicating if this is a dry-run (no permanent modification of the database is done) or
not (permanent modification of the database is done by adding the metadata).}
}
\value{
Invisible TRUE if everything goes correctly.
}
\description{
Collect metadata from yml files
}
\details{
This function will collect the metadata present in the current project (via the \code{metadata.yml} file)
and update the metadata database.
}
\examples{
collect_metadata(con = emf_database, .dry = TRUE)
collect_metadata(con = emf_database, .dry = FALSE)

}
