% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{collect_metadata_creaf}
\alias{collect_metadata_creaf}
\title{Collect metadata for creaf data and models resources}
\usage{
collect_metadata_creaf(con = NULL, .dry = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{\link{collect_metadata}}}
}
\value{
Invisible TRUE if update is needed and performed correctly, invisible FALSE if update is not needed
and process is performed correctly. Metadata \code{yml} object if \code{.dry} i TRUE. Error if the update
failed and there is a mismatch between the updated db tables and the metadata file.
}
\description{
Collect metadata for creaf data and models resources
}
\details{
This function will collect the metadatas present in the selected folder.
}
\examples{
collect_metadata_creaf(con = emf_database, .dry = TRUE)

}
