% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{collect_metadata_external}
\alias{collect_metadata_external}
\title{Collect metadata for external models}
\usage{
collect_metadata_external(external_type, con = NULL, ..., .dry = TRUE)
}
\arguments{
\item{con}{Connection to the metadata database (DBI or pool object)}

\item{...}{arguments passed to \code{\link{external_models_transform}}}

\item{.dry}{Logical indicating if this is a dry-run (no permanent modification of the database is done) or
not (permanent modification of the database is done by adding the metadata).}
}
\value{
Invisible TRUE if update is needed and performed correctly, invisible FALSE if update is not needed
and process is performed correctly. Metadata \code{yml} object if \code{.dry} i TRUE. Error if the update
failed and there is a mismatch between the updated db tables and the metadata file.
}
\description{
Collect metadata for external models
}
\details{
This function will collect the metadata present in the selected excel file.
}
\examples{
collect_metadata_external_models(con = emf_database, .dry = TRUE)

}
