% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{create_model_page}
\alias{create_model_page}
\title{Create a tmodel page in the web project}
\usage{
create_model_page(
  resource_id,
  dest = fs::path(Sys.getenv("WEB_PATH"), "content", "models", resource_id),
  .con = NULL,
  .render_quiet = TRUE,
  .force = FALSE,
  ...
)
}
\arguments{
\item{resource_id}{ID of the workflow resource}

\item{dest}{Path to the resource page tree}

\item{.con}{Connection to the database}

\item{.render_quiet}{Logical, must the render function be quiet?}

\item{.force}{Logical, must the render be done even if not needed?}
}
\value{
invisible TRUE
}
\description{
This function creates a web page for the model resource based on the id
}
\details{
Root web path can be stored in a env var and used here
}
\examples{
create_model_page('test_dummy_model')

}
