% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{create_softwork_page}
\alias{create_softwork_page}
\title{Render pkgdown for softworks}
\usage{
create_softwork_page(
  resource_id,
  dest = fs::path(Sys.getenv("WEB_PATH"), "static", "softworks", resource_id),
  .con = NULL,
  .render_quiet = TRUE,
  .force = FALSE
)
}
\arguments{
\item{resource_id}{Character with the resource ID}

\item{.render_quiet}{Must \code{pkgdown::build_site} be quiet?}
}
\value{
invisible TRUE
}
\description{
Generate the static folder to add to the web project
}
\details{
This function assumes that the softwork repository name is the same as the resource_id
}
\examples{
create_softwork_page('meteospain')

}
