% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{create_tech_doc_page}
\alias{create_tech_doc_page}
\title{Create a technical document page in the web project}
\usage{
create_tech_doc_page(
  resource_id,
  dest = fs::path(Sys.getenv("WEB_PATH"), "content", "tech_docs", resource_id),
  fragment = render_html_fragment(resource_id, .con = .con, .force = .force),
  .con = NULL,
  .render_quiet = TRUE,
  .force = FALSE
)
}
\arguments{
\item{resource_id}{ID of the workflow resource}

\item{dest}{Path to the resource page tree}

\item{fragment}{HTML object with the rendered html fragment of the resource, as obtained by
\code{\link{render_html_fragment}}}

\item{.con}{Connection to the database}

\item{.render_quiet}{Logical, must the render function be quiet?}
}
\value{
invisible TRUE
}
\description{
This function creates a web page for the tech_doc resource based on the id
}
\details{
Tech docs can be Rmd docs or bookdown pages.
}
\examples{
create_tech_doc_page('test_dummy_tech_doc')

}
