% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{create_workflow_page}
\alias{create_workflow_page}
\title{Create a workflow page in the web project}
\usage{
create_workflow_page(
  resource_id,
  dest = fs::path(Sys.getenv("WEB_PATH"), "content", "workflows", resource_id),
  fragment = render_html_fragment(resource_id, .con = .con, .force = .force),
  .con = NULL,
  .render_quiet = TRUE,
  .force = FALSE
)
}
\arguments{
\item{resource_id}{ID of the workflow resource}

\item{dest}{Path to the resource page tree}

\item{fragment}{HTML object with the rendered html fragment of the workflow, as obtained by
\code{\link{render_html_fragment}}}

\item{.con}{Connection to the database}

\item{.render_quiet}{Logical, must the render function be quiet?}
}
\value{
Invisible TRUE
}
\description{
This function creates a web page for the workflow resource based on the id
}
\details{
The path to the web can be stored in a environment variable and this function will take it automatically.
}
\examples{
create_workflow_page('dummy_workflow')

}
