% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{emf_temp_folder}
\alias{emf_temp_folder}
\title{Create temporal folder with unique name}
\usage{
emf_temp_folder()
}
\value{
path to the recently created temporal folder
}
\description{
Allows for multiple temporal folders to be created in the same session
}
\details{
The temporal folder is always the same for an R session, and that can generates problems when using
functions in tests and actions when deferring the temporal folder removal. To solve this we create temporal
folders with the tempfile name, create the folder and return the path
}
\examples{
temp_folder <- emf_temp_folder()
fs::dir_exists(temp_folder)

}
