% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfdata_fileutils.R
\name{readRDS_emfdata}
\alias{readRDS_emfdata}
\alias{curl_emfdata}
\alias{saveRDS_emfdata}
\alias{list_emfdata}
\alias{upload_emfdata}
\alias{emfdata_file_utilities}
\alias{download_emfdata}
\title{File utilities for EMF_DATA}
\usage{
readRDS_emfdata(emfdata_base, file)

curl_emfdata(emfdata_base, file)

saveRDS_emfdata(obj, emfdata_base, file)

list_emfdata(emfdata_base, dir = "")

upload_emfdata(local_file, emfdata_base, file)

download_emfdata(emfdata_base, file)
}
\arguments{
\item{emfdata_base}{base path in EMF_DATA}

\item{file}{target file in EMF_DATA}

\item{obj}{object to be written as RDS file}

\item{dir}{directory of EMF_DATA to be listed}

\item{local_file}{Path to local file to be read or written}
}
\description{
File utilities for EMF_DATA
}
