% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_utils.R
\name{matomo_lfc_events}
\alias{matomo_lfc_events}
\title{Event stats for LFC apps}
\usage{
matomo_lfc_events(last = 31, period = "day")
}
\arguments{
\item{last}{number of last records to retrieve}

\item{period}{period to retrieve (day, week, month...)}
}
\description{
Connect to matomo API and retrieve LFC apps event stats
}
\details{
This function assumes some environmental variables are available
}
