% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{metadata_db_con}
\alias{metadata_db_con}
\title{Connect to the EMF metadata database}
\usage{
metadata_db_con(
  host = Sys.getenv("EMF_DATABASE_HOST"),
  user = Sys.getenv("EMF_DATABASE_USER"),
  password = Sys.getenv("EMF_DATABASE_PASS"),
  dbname = Sys.getenv("EMF_DATABASE"),
  .envir = parent.frame()
)
}
\arguments{
\item{host}{db host address/name to resolve}

\item{user}{db user}

\item{password}{db user password}

\item{dbname}{db name to connect}

\item{.envir}{environment in which the defer call takes place}
}
\value{
A \code{pool} object with the database connection
}
\description{
Pool connection to the EMF metadata database
}
\details{
This function returns a pool object with a connection to the postgreSQL
metadata database. It also defers (\code{\link[withr]{defer}}) the pool
closing to the parent environment by default.
The connection params default to environment variables, set these with
\code{Sys.setenv} or \code{withr::with_envvar}, or with an \code{.Renviron}
file.
}
