% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{nodes_diagram_generator}
\alias{nodes_diagram_generator}
\title{Generate resource nodes graph plot.}
\usage{
nodes_diagram_generator(.con = NULL)
}
\arguments{
\item{.con}{metadata database connetion}
}
\value{
An echarts4r object (htmlwidget)
}
\description{
Generate resource nodes graph plot, indicating the relationships between them.
}
\examples{
nodes_diagram_generator()

}
