% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{pq__text_to_vector_parser}
\alias{pq__text_to_vector_parser}
\title{Parsing "pq__text" columns to flattened character vectors.}
\usage{
pq__text_to_vector_parser(pq__text)
}
\arguments{
\item{pq__text}{character string as obtained from public_* metadata db tables}
}
\value{
A character vector with the elements of the metadata field
}
\description{
Parsed the defined format to obtain the array columns from postgres
}
\details{
We have defined a format: \code{\\{element1,element2,...,elementn\\}} to obtain the array postgres values
when importing to R tibbles. This functions parses this and transform it to a vector to use in yamls or
others.
}
\examples{
pq__text_to_vector_parser("{dummy,workflow,larara}")

}
