% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{public_tech_docs}
\alias{public_tech_docs}
\title{Retrieve the public tech_docs table.}
\usage{
public_tech_docs(...)
}
\arguments{
\item{...}{arguments for \code{\link{use_public_tables}} except for \code{resource_type}. If the former is
specified and error is raised.}
}
\value{
A tibble with the public resources table queried
}
\description{
Retrieve the public tech_docs table.
}
\examples{
use_public_table('tech_docs')

}
