% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfdata_fileutils.R
\name{read_emfdata}
\alias{read_emfdata}
\alias{contents_emfdata}
\alias{save_emfdata}
\title{Access data from EMF data server}
\usage{
read_emfdata(
  data_path,
  file,
  ftps_address_port = Sys.getenv("emfdata_ftps_address_port"),
  userpwd = Sys.getenv("emfdata_userpwd"),
  download_tmp = FALSE,
  noread = FALSE,
  .fun = NULL,
  ...,
  .env = parent.frame()
)

contents_emfdata(
  data_path,
  ftps_address_port = Sys.getenv("emfdata_ftps_address_port"),
  userpwd = Sys.getenv("emfdata_userpwd")
)

save_emfdata(
  x,
  data_path,
  output_file,
  ftps_address_port = Sys.getenv("emfdata_ftps_address_port"),
  userpwd = Sys.getenv("emfdata_userpwd")
)
}
\arguments{
\item{data_path}{base path in EMF data server}

\item{file}{file name (with extension) to read}

\item{userpwd}{User and password (in the form "user:pasword"). Defaults to
\code{Sys.getenv("emfdata_userpwd")}.}

\item{download_tmp}{Logical indicating if the file should be downloaded in a temporal file
before applying \code{.fun} instead of being treated as a curl connection by \code{.fun}.
Defaults to \code{FALSE}.}

\item{noread}{Logical indicating if the file should be downloaded only and no attempt to read it
should be done. Defaults to \code{FALSE}. Only used if \code{download_tmp} is TRUE.}

\item{.fun}{Function to read the connection (or file if \code{download_tmp}) is TRUE. If no
function is supplied, \code{read_emfdata} will try default ones based on the file extension.
For a comprehensible list of default functions see Details.}

\item{...}{Extra arguments for \code{.fun}}

\item{.env}{Environment to load the data in, if loading RData files}

\item{x}{\code{x} can be either an object or a character vector of length one with a path to a
local file. If \code{x} is an object, then it will be saved as an \code{rds} file with
\code{\link[base]{saveRDS}}. If \code{x} is a path, then the file is uploaded as is.}

\item{output_file}{character indicating the name of the output file in the EMF data server}

\item{ftps_addres_port}{Address and port (in the form "protocol://address:port/"). Defaults to
\code{Sys.getenv("emfdata_ftps_address_port")}.}
}
\description{
Read files in the EMF data server
}
\details{
\code{read_emfdata} can automatically detect the type of file and apply the default function
for that file. Using curl connections or downloading the file is indicated by the
\code{download_tmp} argument, by default FALSE. Extra arguments for reading function can be
supplied.
}
\section{Functions}{
\itemize{
\item \code{contents_emfdata()}: List available contents for the provided data path

\item \code{save_emfdata()}: Save files to the EMF data server

}}
\section{Default functions}{

If no \code{.fun} is supplied, \code{read_emfdata} will try to match a default function with
the file extension. If no match is found, the connection (or the file if \code{download_tmp}
is TRUE) are returned for the user to deal with.
}

\section{RData files}{

RData files are loaded with \code{\link[base]{load}}. This means that objects contained in the
RData file will be loaded by default into the environment \code{read_emfdata} is called. The
names of the objects loaded are returned in a character vector.
}

