% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{render_html_fragment}
\alias{render_html_fragment}
\title{Render resource Rmd as html fragment}
\usage{
render_html_fragment(
  resource_id,
  .render_quiet = TRUE,
  .force = FALSE,
  .con = NULL
)
}
\arguments{
\item{resource_id}{Character with the resource ID}

\item{.render_quiet}{Must \code{rmarkdown::render} be quiet?}

\item{.force}{Should the fragment be rendered even if there is no new commit in resource repository?}

\item{.con}{Connection to pass, if provided}
}
\value{
html object
}
\description{
Generate an html fragment with the contents of the rendered Rmd file.
}
\details{
This function assumes that the Rmd file is called the same as the resource id
}
\examples{
render_html_fragment('test_dummy_workflow')

}
