% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_web_functions.R
\name{render_resource_pages}
\alias{render_resource_pages}
\title{Render all resource pages}
\usage{
render_resource_pages(
  ...,
  .con = NULL,
  .force = FALSE,
  .web_path = Sys.getenv("WEB_PATH")
)
}
\arguments{
\item{...}{Arguments passed to render_resource_type}

\item{.con}{Connection to the metadata database}
}
\value{
A named vector for resources indicating (TRUE or FALSE) if they have
been updated and rendered
}
\description{
Render all resource pages if needed
}
\details{
This function starts the process to update and render all pages for resources
present in the metadata database.
}
\examples{

render_resource_pages()

}
