% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{render_workflow_fragment}
\alias{render_workflow_fragment}
\title{Render workflow as html fragment}
\usage{
render_workflow_fragment(resource_id, .envir = parent.frame())
}
\arguments{
\item{resource_id}{Character with the resource ID}
}
\value{
html object
}
\description{
Generate an html fragment with the contents of the rendered workflow
}
\examples{
render_workflow_fragment('dummy_workflow')

}
