% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{resource_color}
\alias{resource_color}
\title{Return EMF resources color codes}
\usage{
resource_color(resource = NULL)
}
\arguments{
\item{resource}{Resource name from which get the color hex code.}
}
\value{
If \code{resource} is NULL (default), a tibble with two columns, resource_name and hex_color_code.
If \code{resource} is not NULL, function returns directly the hex color code for the resource specified.
}
\description{
Get the color hex codes for the different EMF resources
}
\details{
Each EMF resource has a color assigned, this function returns the codes
}
\examples{
resource_color()
resource_color('model')

}
