% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{update_all_resource_pages}
\alias{update_all_resource_pages}
\title{Update all resource pages}
\usage{
update_all_resource_pages(...)
}
\arguments{
\item{...}{arguments to pass to the create_*_page functions.}
}
\value{
a list of named lists with the resources and if they were updated or not (logical)
}
\description{
Update all resources pages if needed
}
\details{
This function update all resources pages that need it, or all if \code{.force = TRUE})
}
\examples{
update_all_resource_pages()

}
