% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops_helpers.R
\name{update_docs_site}
\alias{update_docs_site}
\title{Update Documentation site}
\usage{
update_docs_site(dest)
}
\arguments{
\item{dest}{path to the production folder}
}
\value{
invisible TRUE if all goes well, error if not
}
\description{
This function retrieves the latest github version of the docs and update the
doc server
}
\details{
This function is intended to be used in the doc server, it has not
remote option.
}
