% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_web_functions.R
\name{update_emf_web}
\alias{update_emf_web}
\title{Update web repository}
\usage{
update_emf_web(
  repo = "emf_web",
  org = "emf-creaf",
  commit_message = glue::glue("{Sys.time()} automatic update"),
  github_pat = Sys.getenv("GITHUB_PAT"),
  remote = TRUE,
  prod_folder = Sys.getenv("PROD_FOLDER"),
  prod_host = Sys.getenv("PROD_HOST"),
  prod_pass = Sys.getenv("PROD_PASS"),
  .con = NULL,
  .force = FALSE,
  .dry_push = FALSE
)
}
\arguments{
\item{repo}{character with repo name}

\item{org}{character with organization/user name}

\item{commit_message}{Character with the commit message}

\item{github_pat}{Character wiht the remote token (GitHub)}

\item{remote}{logical indicating if executing from outside prod server,
default to TRUE}

\item{.con}{pool::pool object with the metadata database connection info}

\item{.force}{logical indicating if the render shoud be forced even if
everything is up-to-date}

\item{.dry_push}{logical to pass to  \code{\link{commit_push_web_repo}}}
}
\value{
invisible FALSE if no changes are to be commited and prod is not
updated (unless \code{.force = TRUE}). TRUE if everything goes correctly.
}
\description{
Update web repository with new/modified content
}
\details{
This function clone the latest commit from the remote web repository (GitHub)
and render all pages that need update (\code{\link{render_resource_pages}}).
After that if any changes have been made, it updates the web repository with
the changes. After updating the repository, production server is updated.
}
