% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops_helpers.R
\name{update_metadata_collection_ga}
\alias{update_metadata_collection_ga}
\title{Update metadata collection GA for all resources}
\usage{
update_metadata_collection_ga(
  .con = NULL,
  .dry_push = TRUE,
  .github_path = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{.con}{metadata database connection to retrieve resources info}

\item{.dry_push}{logical indicating if commit changes must be hold from pushing}

\item{.github_pat}{GitHub PAT}
}
\value{
Named list with the resources and the update status
}
\description{
Update metadata collection GA for all resources
}
\details{
This function takes all resources repositories and update the collecing
metadata GA with the newer version from \code{\link{EMFtoolbox}}. Each
resource is served with the corresponding metadata GA file (normal for
all resources except for models and data, external or creaf, that
receives the corresponding GA for the type).
}
