% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{update_resource_pages_by_type}
\alias{update_resource_pages_by_type}
\title{Update pages for resources}
\usage{
update_resource_pages_by_type(
  emf_type = c("workflow", "tech_doc", "model", "data", "softwork"),
  resources = NULL,
  ...
)
}
\arguments{
\item{emf_type}{Character with the type of resources to update}

\item{resources}{Character vector with the resource IDs to update. Default to NULL, which retrieves all
public workflows and update them}

\item{...}{Arguments for \code{\link{create_*_page}}, except for resource_id.}
}
\value{
a named list with the resources and if they were updated or not (logical)
}
\description{
Update pages for resources
}
\details{
This function takes a resource type and a resource names list and create the pages if needed.
If no resources list is supplied, all resources of the selected type are updated.
}
\examples{
update_resource_pages_by_type('workflow', c('test_dummy_workflow', 'non_existent_workflow'))

}
