% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{update_workflow_pages}
\alias{update_workflow_pages}
\title{Update pages for workflows}
\usage{
update_workflow_pages(resources = NULL, ...)
}
\arguments{
\item{resources}{Character vector with the resource IDs to update. Default to NULL, which retrieves all
public workflows and update them}

\item{...}{Arguments for \code{\link{create_workflow_page}}, except for resource_id.}
}
\value{
a named list with the resources and if they were updated or not (logical)
}
\description{
Update pages for workflows
}
\details{
This function is a vectorised and safe (purrr::possibly) version of \code{\link{create_workflow_page}}, and
hence, it depends on a connection to the database (created from environment variables).
}
\examples{
update_workflow_pages(c('test_dummy_workflow', 'non_existent_workflow'))

}
