% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_functions.R
\name{use_emf_ga}
\alias{use_emf_ga}
\title{Use EMF github actions}
\usage{
use_emf_ga(.template)
}
\arguments{
\item{.template}{Template file name (with extension) to use}
}
\value{
invisible TRUE if success
}
\description{
Create the file needed to use the desired EMF GA
}
\details{
This function allows to select an EMF github action template (inst/ga_templates) to use in the resource
project
}
\examples{
use_emf_ga('collect_metadata.yml')

}
