% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{use_metadata_yml}
\alias{use_metadata_yml}
\title{Create and modify metadata files content}
\usage{
use_metadata_yml(..., .template = NULL)
}
\arguments{
\item{...}{Name-value pairs for metadata fields to modify or add to the metadata file}

\item{.template}{Template file to use}
}
\value{
Metadata content ready to copy&paste to the \code{metadata.yml} file printed in the
console and returned as a \code{yml} object.
}
\description{
Create and modify the content for EMF metadata files. Internal templates can be
used.
}
\details{
The purpose of this function is to automatically create the contents of a metadata file for
any EMF resource
}
\examples{
# TODO

}
