% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{use_public_table}
\alias{use_public_table}
\title{Access the public_* tables.}
\usage{
use_public_table(
  resource_type = c("workflows", "tech_docs", "models", "data", "softworks", "all"),
  ...,
  .con = NULL
)
}
\arguments{
\item{resource_type}{Character indicating the type of resource to query}

\item{...}{arguments for \code{dplyr::filter} to filter the results}

\item{.con}{Connection to the database (pool). NULL by default, connecting with the environment variables.}
}
\value{
A tibble with the public resources table queried
}
\description{
Access to the public resources tables
}
\examples{
use_public_table('workflows')

}
