% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{create_from_emf_github}
\alias{create_from_emf_github}
\title{Clone and set project from github}
\usage{
create_from_emf_github(
  resource_id,
  .envir = parent.frame(),
  .con = NULL,
  .external = FALSE,
  .external_id
)
}
\arguments{
\item{resource_id}{Resource ID}

\item{.envir}{envir for \code{withr::defer}, default to \code{parent.frame()}}

\item{.con}{connection to the db}

\item{.external}{is the resource external}

\item{.external_id}{if .external is TRUE, the id of the external data/models repository}
}
\description{
Temporal cloned repository to work with the resource (render...)
}
\details{
This function creates a temporal folder, clone the desired repository and set the active project on it
}
