% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_table}
\alias{metadata_table}
\title{Metadata definitions table.}
\format{
A data frame with 4 variables:
\describe{
\item{Field}{ID of the metadata field}
\item{Scopes}{Reources the metadata is intended for}
\item{Categories}{Metadata categories}
\item{Definition}{Metadata field definition}
}
}
\usage{
metadata_table
}
\description{
A dataframe containing the metadata fields available in the EMFverse.
}
\keyword{datasets}
