% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{render_html_fragment}
\alias{render_html_fragment}
\title{Render resource Rmd as html fragment}
\usage{
render_html_fragment(
  resource_id,
  .envir = parent.frame(),
  .render_quiet = TRUE
)
}
\arguments{
\item{resource_id}{Character with the resource ID}

\item{.envir}{envir for \code{withr::defer}, default to \code{parent.frame()}}

\item{.render_quiet}{Must \code{rmarkdown::render} be quiet?}
}
\value{
html object
}
\description{
Generate an html fragment with the contents of the rendered Rmd file.
}
\details{
This function assumes that the Rmd file is called the same as the resource id
}
\examples{
render_html_fragment('test_dummy_workflow')

}
