% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{use_metadata_yml}
\alias{use_metadata_yml}
\alias{use_workflow_yml}
\alias{use_tech_doc_yml}
\alias{use_data_yml}
\alias{use_model_yml}
\alias{use_softwork_yml}
\title{Create and modify metadata files content}
\usage{
use_metadata_yml(
  ...,
  .template = system.file("metadata_templates", "core.yml", package = "EMFtoolbox"),
  .write = FALSE
)

use_workflow_yml(..., .write = FALSE)

use_tech_doc_yml(..., .write = FALSE)

use_data_yml(..., .write = FALSE)

use_model_yml(..., .write = FALSE)

use_softwork_yml(..., .write = FALSE)
}
\arguments{
\item{...}{Name-value pairs for metadata fields to modify or add to the metadata file}

\item{.template}{Template file to use}

\item{.write}{Logical indicating if the \code{metadata.yml} file must be written (TRUE) or not (FALSE)}
}
\value{
Metadata content ready to copy&paste to the \code{metadata.yml} file printed in the
console and returned as a \code{yml} object.
}
\description{
Create and modify the content for EMF metadata files. Internal templates can be
used.
}
\details{
The purpose of these functions is to automatically create the contents for the metadata file of any EMF
resource
}
\section{Functions}{
\itemize{
\item \code{use_workflow_yml}: Equivalent to using \code{use_metadata_yml} with
\code{.template = system.file('metadata_templates', 'workflow.yml', package = 'EMFtoolbox')}

\item \code{use_tech_doc_yml}: Equivalent to using \code{use_metadata_yml} with
\code{.template = system.file('metadata_templates', 'tech_doc.yml', package = 'EMFtoolbox')}

\item \code{use_data_yml}: Equivalent to using \code{use_metadata_yml} with
\code{.template = system.file('metadata_templates', 'data.yml', package = 'EMFtoolbox')}

\item \code{use_model_yml}: Equivalent to using \code{use_metadata_yml} with
\code{.template = system.file('metadata_templates', 'model.yml', package = 'EMFtoolbox')}

\item \code{use_softwork_yml}: Equivalent to using \code{use_metadata_yml} with
\code{.template = system.file('metadata_templates', 'softwork.yml', package = 'EMFtoolbox')}
}}

\section{Templates}{

The easiest way to create the metadata fields for any EMF resource is using the corresponding
\code{use_*_yml} function. But in specials cases can be useful to provide a custom template. This
template must be a \code{yml} file with the required metadata fields. Check \code{EMFtoolbox}
predetermined templates, located in the package installation folder, to see examples.
}

\examples{
use_metadata_yml()
use_metadata_yml(emf_type = 'workflow', tags = c('dummy', 'foo'))
use_workflow_yml(tags = c('dummy', 'foo'))
use_data_yml(tags = c('bar', 'baz'))
use_model_yml(nodes = c('bar', 'dummy'))

}
